# services/events-service/python-flask-server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Events API for ZeroWing Event Management System."
  version: "1.0.0"
  title: "Events API"
host: "localhost:8081"
basePath: "/v1"
schemes:
- "https"
- "http"
paths:
...
  /reviews:
    get:
      tags:
      - "Reviews"
      summary: "Retrieve a List of Reviews. Search by Event ID and Interaction Account\
        \ ID."
      operationId: "list_reviews"
      produces:
      - "application/json"
      parameters:
      - name: "event_id"
        in: "query"
        description: "The Event to search."
        required: false
        type: "string"
      - name: "interaction_acount_id"
        in: "query"
        description: "The ID of the user who is reading the review."
        required: false
        type: "string"
      responses:
        "200":
          description: "Returns a list of Reviews. Returns empty array if no Reviews\
            \ match the search criteria."
          schema:
            $ref: "#/definitions/ReviewList"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.reviews_controller"
    post:
      tags:
      - "Reviews"
      summary: "Used to create a Review."
      operationId: "create_review"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Review object containing the Review details."
        required: true
        schema:
          $ref: "#/definitions/Review"
      responses:
        "201":
          description: "Review created."
          schema:
            $ref: "#/definitions/Review"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.reviews_controller"
  /reviews/{review_id}:
    put:
      tags:
      - "Reviews"
      summary: "Used to update a Review. Replaces the Review resource."
      operationId: "update_review"
      consumes:
      - "application/json"
      parameters:
      - name: "review_id"
        in: "path"
        description: "ID of the Review to be updated."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Review object to update."
        required: true
        schema:
          $ref: "#/definitions/Review"
      responses:
        "200":
          description: "Review successfully updated."
          schema:
            $ref: "#/definitions/Review"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "404":
          description: "Review not found."
          schema:
            $ref: "#/definitions/ReviewNotFoundError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.reviews_controller"
  /review_interaction:
    post:
      tags:
      - "Reviews"
      summary: "Used to create a Review Intercation record."
      operationId: "create_review_interaction"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Review Interaction object containing the Review Interaction\
          \ details."
        required: true
        schema:
          $ref: "#/definitions/ReviewInteraction"
      responses:
        "201":
          description: "Review Interaction Created."
          schema:
            $ref: "#/definitions/ReviewInteraction"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.reviews_controller"
  /review_interaction/{interaction_id}:
    put:
      tags:
      - "Reviews"
      summary: "Used to update a Review Interaction record."
      operationId: "update_review_interaction"
      consumes:
      - "application/json"
      parameters:
      - name: "interaction_id"
        in: "path"
        description: "ID of the Review Interaction record to be updated."
        required: true
        type: "integer"
        format: "int64"
      - in: "body"
        name: "body"
        description: "Review Interaction object to update."
        required: true
        schema:
          $ref: "#/definitions/ReviewInteraction"
      responses:
        "200":
          description: "Review successfully updated."
          schema:
            $ref: "#/definitions/ReviewInteraction"
        "400":
          description: "Invalid input."
          schema:
            $ref: "#/definitions/InvalidInputError"
        "404":
          description: "Review Interaction not found."
          schema:
            $ref: "#/definitions/ReviewInteractionNotFoundError"
        "500":
          description: "Unexpected failure in the service layer."
          schema:
            $ref: "#/definitions/UnexpectedServiceError"
      x-swagger-router-controller: "swagger_server.controllers.reviews_controller"
...
definitions:
...
  ReviewList:
    type: "array"
    items:
      $ref: "#/definitions/Review"
...
  Review:
    type: "object"
    properties:
      review_id:
        type: "integer"
        format: "int64"
        example: 1
      event_id:
        type: "integer"
        format: "int64"
        example: 3
      reviewer_account_id:
        type: "integer"
        format: "int64"
        example: 3
      upvotes:
        type: "integer"
        format: "int64"
        example: 20
      rating:
        type: "integer"
        format: "int64"
        example: 5
      flag_count:
        type: "integer"
        format: "int64"
        example: 2
      review_text:
        type: "string"
        example: "We had a great time. Fantastic Event. Highly recommended."
      reply_text:
        type: "string"
        example: "Glad you had a great time."
      review_timestamp:
        type: "string"
        example: "2015-03-25T19:00:00+10:00"
      review_status:
        type: "string"
        enum:
        - "Active"
        - "Removed"
      review_interaction:
        type: "object"
        example:
          interaction_id: 1
          review_id: 1
          interaction_account_id: 2
          review_upvoted: true
          review_flagged: false
        allOf:
        - $ref: "#/definitions/ReviewInteraction"
    example:
      review_id: 1
      review_timestamp: "2015-03-25T19:00:00+10:00"
      event_id: 3
      upvotes: 20
      flag_count: 2
      reviewer_account_id: 3
      reply_text: "Glad you had a great time."
      rating: 5
      review_text: "We had a great time. Fantastic Event. Highly recommended."
      review_status: "Active"
      review_interaction:
        interaction_id: 1
        review_id: 1
        interaction_account_id: 2
        review_upvoted: true
        review_flagged: false
  ReviewInteraction:
    type: "object"
    properties:
      interaction_id:
        type: "integer"
        format: "int64"
        example: 1
      review_id:
        type: "integer"
        format: "int64"
        example: 2
      interaction_account_id:
        type: "integer"
        format: "int64"
        example: 2
      review_upvoted:
        type: "boolean"
        example: true
      review_flagged:
        type: "boolean"
        example: true
    example:
      review_upvoted: true
      review_id: 2
      interaction_account_id: 2
      interaction_id: 1
      review_flagged: true
...
  Error:
    type: "object"
    properties:
      code:
        type: "string"
      type:
        type: "string"
      message:
        type: "string"
...
  ReviewNotFoundError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 404
      type: "ReviewNotFoundError"
      message: "The following Review does not exist: %review_id%"
  ReviewInteractionNotFoundError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 404
      type: "ReviewInteractionNotFoundError"
      message: "The following Review Interaction does not exist: %review_interaction_id%"
...
  InvalidInputError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 400
      type: "InvalidInputError"
      message: "The following mandatory fields were not provided: %field_name_1%,\
        \ %field_name_2%"
  UnexpectedServiceError:
    allOf:
    - $ref: "#/definitions/Error"
    - {}
    example:
      code: 500
      type: "UnexpectedServiceError"
      message: "The following unexpected runtime error has occurred: %error%"